requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
sqlalchemy>=2.0.32
pg8000>=1.31.0
requests>=2.32.3
mysql-connector-python>=9.0.0
typing-extensions>=4.12.2
google-generativeai>=0.5.0
python-telegram-bot>=13.15